Update scatterplot instructions
It's too late to add this to the students notebook, so it is more for future years. I think this description better capture what we have in the solution (which is better than creating separate plots, since the axes scales won't be the same and a direct comparison is more difficult).